+++ app/Services/Ftp.php
@@ -335,7 +335,7 @@ class Ftp extends Directory_Listing_Base implements Service {
 				),
 				array(
 					'action' => 'efml_save_as_directory( "ftp", actualDirectoryPath, login, password, "", config.term );',
-					'label'  => __( 'Save active directory as directory archive', 'external-files-in-media-library' ),
+					'label'  => __( 'Save active directory as your external source', 'external-files-in-media-library' ),
 				),
 			)
 		);