+++ app/Services/Rest.php
@@ -374,7 +374,7 @@ class Rest extends Directory_Listing_Base implements Service {
 				),
 				array(
 					'action' => 'efml_save_as_directory( "' . $this->get_name() . '", actualDirectoryPath, login, password, "" );',
-					'label'  => __( 'Save active directory as directory archive', 'external-files-in-media-library' ),
+					'label'  => __( 'Save active directory as your external source', 'external-files-in-media-library' ),
 				),
 			)
 		);