+++ app/Services/Youtube.php
@@ -540,7 +540,7 @@ class Youtube extends Directory_Listing_Base implements Service {
 				),
 				array(
 					'action' => 'efml_save_as_directory( "youtube", actualDirectoryPath, url, "", apiKey );',
-					'label'  => __( 'Save active directory as directory archive', 'external-files-in-media-library' ),
+					'label'  => __( 'Save active directory as your external source', 'external-files-in-media-library' ),
 				),
 			)
 		);