+++ changelog.md
@@ -25,6 +25,7 @@
 
 ### Changed
 
+- Renamed "Directory Archive" to "Your external sources"
 - Show processed file URLs during manual started synchronization
 - Hide import button for unsupported files in directory archive
 - Small optimizations on multiple codes